% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateYoungBiomasses.R
\name{spinUpPartial}
\alias{spinUpPartial}
\title{Update Biomass using a spinup type approach}
\usage{
spinUpPartial(
  pixelCohortData,
  speciesEcoregion,
  maxAge,
  minRelativeB,
  species,
  sppEquiv,
  sppEquivCol,
  sppColorsVect,
  paths,
  currentModule,
  modules
)
}
\arguments{
\item{pixelCohortData}{A pixelCohortData object (must have columns pixelIndex,
ecoregionGroup, age, and B). Should only have rows where the
age <= maxAge.
It should have none where age = 0}

\item{speciesEcoregion}{A speciesEcoregion object (must have ecoregionGroup, maxB, maxANPP)}

\item{maxAge}{A numeric scalar, indicating the maximum age to simulation biomass.
This should correspond to the maximum age in pixelCohortData}

\item{minRelativeB}{A minRelativeB object}

\item{species}{A species table with species-level parameters}

\item{sppColorsVect}{A sppColorsVect object (required, but not used)}

\item{paths}{A list of spades paths e.g., from paths(sim)}

\item{currentModule}{A character string of the current module e.g., from currentModule(sim)}

\item{modules}{A list of character strings of the modules in the sim, e.g., from modules(sim)}
}
\description{
This function takes a pixelCohortData, and updates all the B values according to
the succession and growth dynamics provided by \code{species} and \code{speciesEcoregion}
objects. It does this by setting B to zero in each pixelIndex, growing each pixelIndex
until it hits the age given by age, for each pixelIndex independently. This uses
Biomass_core module to do this. If this module does not exist in the modules(sim),
then it will download it and use the latest development branch version of Biomass_core.
}
