% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateYoungBiomasses.R
\name{updateYoungBiomasses}
\alias{updateYoungBiomasses}
\title{Update young biomasses}
\usage{
updateYoungBiomasses(young, modelBiomass, ...)
}
\arguments{
\item{young}{data.table not unlinke \code{cohortData}}

\item{modelBiomass}{named list with items "mod", "pred", "rsq", "scaledVarsModelB".}

\item{...}{For anything, used for Cache. Not used internally here.}
}
\description{
Update young biomasses
}
